// src/Modules/Shared/Interfaces/AccountCreationData.ts
export interface AccountCreationData {
  firstName: string;
  lastName: string;
  phoneNumber: string;
  email: string;
  newPassword: string;
  address: string;
  city: string;
  state: string;
  dob: string;
  insuranceName: string;
  policyHolderName: string;
  memberId: string;
  memberDOB: string;
  insuranceType: string;
  groupNumber: string;
  effectiveDate: string;
  relToPolicyHolder: string;
};